ref: move prepareEmailSubmission into ts
