+++ tests/unit/backend/controllers/email-submissions.server.controller.spec.js
@@ -10,7 +10,7 @@ const dbHandler = require('../helpers/db-handler')
 const { ObjectID } = require('bson-ext')
 const MailService = require('../../../../dist/backend/app/services/mail/mail.service')
   .default
-
+const EmailSubmissionsMiddleware = require('../../../../dist/backend/app/modules/submission/email-submission/email-submission.middlewares')
 const User = dbHandler.makeModel('user.server.model', 'User')
 const Agency = dbHandler.makeModel('agency.server.model', 'Agency')
 const Form = dbHandler.makeModel('form.server.model', 'Form')
@@ -733,7 +733,7 @@ describe('Email Submissions Controller', () => {
           controller.validateEmailSubmission,
           submissionsController.injectAutoReplyInfo,
           spcpController.appendVerifiedSPCPResponses,
-          controller.prepareEmailSubmission,
+          EmailSubmissionsMiddleware.prepareEmailSubmission,
           sendSubmissionBack,
         )
     })
